requirements.txt
# PDF parsing
pymupdf>=1.24.0

# Embeddings (local, no API key needed)
sentence-transformers>=3.0.0

# Vector store (local persistent)
chromadb>=0.5.0

# LLM + MCP
openai>=1.30.0
mcp[cli]>=1.0.0

# Utilities
python-dotenv>=1.0.0
